+++ app/_common/hooks/queries/useQuerySignUpUser.ts
@@ -2,14 +2,10 @@ import { useQuery } from "@tanstack/react-query";
 
 import { getSignUpUser } from "@/app/_common/api/auth";
 
-import { SignUpUser } from "../_type/signup";
-
 export const useQuerySignUpUser = (hasAccessToken = true) => {
-  const { data, isLoading } = useQuery<SignUpUser>({
+  return useQuery({
     queryFn: getSignUpUser,
     queryKey: ["user"],
     enabled: hasAccessToken,
   });
-
-  return { data, isLoading };
 };
