+++ app/my-page/_components/MyProjectList.tsx
@@ -2,13 +2,14 @@ import Link from "next/link";
 
 import Icon from "@/app/_common/components/Icon";
 
-import { useQueryUserData } from "../_hooks/useQueryUserData";
+import { useQuerySignUpUser } from "@/app/_common/hooks/queries/useQuerySignUpUser";
+
 import { useQueryProjectList } from "../_hooks/useQueryProjectList";
 import ProjectCardSkeleton from "./ProjectCardSkeleton";
 import ProjectCard from "./ProjectCard";
 
 function MyProjectList() {
-  const { data: userData } = useQueryUserData();
+  const { data: userData } = useQuerySignUpUser();
   const { data } = useQueryProjectList(userData?.id, 3);
 
   return (
